votes left
